cloud/requirements.txt
azure-cognitiveservices-speech
azure-ai-textanalytics
azure-core
matplotlib
numpy
nltk
flask
python-dotenv
azure-cosmos


Werkzeug
Flask-JWT-Extended
PyJWT 
pydub
google-auth
google-auth-oauthlib
google-auth-httplib2
requests
bcrypt
azure-storage-blob
celery
redis
gunicorn
pycryptodome
azure-identity
azure-keyvault-secrets